+++ packages/veui-theme-dux/components/Breadcrumb.js
@@ -1,8 +1,0 @@
-import '../icons/angle-double-right'
-import config from 'veui/managers/config'
-
-config.defaults({
-  'breadcrumb.icons': {
-    next: 'angle-double-right'
-  }
-})