+++ packages/veui-theme-dux/components/Button.js
@@ -1,8 +1,0 @@
-import '../icons/loading'
-import config from 'veui/managers/config'
-
-config.defaults({
-  'button.icons': {
-    loading: 'loading'
-  }
-})