+++ packages/veui-theme-dux/components/Dialog.js
@@ -1,9 +1,0 @@
-import '../icons/cross'
-
-import config from 'veui/managers/config'
-
-config.defaults({
-  'dialog.icons': {
-    close: 'cross'
-  }
-})