+++ packages/veui-theme-dux/components/Field.js
@@ -1,8 +1,0 @@
-import '../icons/exclamation-circle'
-import config from 'veui/managers/config'
-
-config.defaults({
-  'field.icons': {
-    alert: 'exclamation-circle'
-  }
-})