+++ packages/veui-theme-dux/components/Option.js
@@ -1,8 +1,0 @@
-import '../icons/check'
-import config from 'veui/managers/config'
-
-config.defaults({
-  'option.icons': {
-    checked: 'check'
-  }
-})