+++ packages/veui-theme-dux/components/Progress.js
@@ -1,8 +1,0 @@
-import '../icons/check'
-import config from 'veui/managers/config'
-
-config.defaults({
-  'progress.icons': {
-    success: 'check'
-  }
-})